pyproject.toml
[tool.poetry]
name = "pyedahelper"
version = "0.1.0"
description = "Python package that simplifies EDA."
authors = ["Ofer Mansour <mansour.ofer@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.1"
altair = "^4.0.1"
statistics = "^1.0.3"
seaborn = "^0.10.0"
matplotlib = "^3.2.0"
numpy = "^1.18.1"
scipy = "^1.4.1"

[tool.poetry.dev-dependencies]
sphinx = "^2.4.3"
sphinxcontrib-napoleon = "^0.7"
pytest-cov = "^2.8.1"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
